Improve SuperTextField tests (Resolves #615)  - replaced existing keyboard simulation with flutter_test_robots version  - removed SuperTextField test tools from public API and started reworking into robot and interactor
